requirements.txt
asgiref==3.8.1
Django==4.2
sqlparse==0.5.0

# Django REST Framework
djangorestframework==3.14.0

# OAuth 2.0
django-oauth-toolkit==2.3.0

# CORS (para permitir peticiones desde React)
django-cors-headers==4.3.1

# Opcional pero recomendado
python-decouple==3.8